requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
firebase-admin==6.2.0
requests==2.31.0
python-dotenv==1.0.0
oauthlib==3.2.2
requests-oauthlib==1.3.1
gunicorn==21.2.0
